Move AWS Batch validators code from utils to awsbatch_validators
Moved related tests from test_utils to test_awsbatch_validators and adapted to use AWSApi
