+++ cli/tests/pcluster/validators/test_awsbatch_validators.py
@@ -8,18 +8,33 @@
 # or in the "LICENSE.txt" file accompanying this file. This file is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES
 # OR CONDITIONS OF ANY KIND, express or implied. See the License for the specific language governing permissions and
 # limitations under the License.
+import logging
+from itertools import product
+from re import escape
+
 import pytest
 from assertpy import assert_that
+from botocore.exceptions import ClientError, EndpointConnectionError
 
+import pcluster.validators.awsbatch_validators as awsbatch_validators
 from pcluster.utils import InstanceTypeInfo
 from pcluster.validators.awsbatch_validators import (
     AwsbatchComputeInstanceTypeValidator,
     AwsbatchComputeResourceSizeValidator,
     AwsbatchInstancesArchitectureCompatibilityValidator,
     AwsbatchRegionValidator,
+    BatchErrorMessageParsingException,
 )
+from tests.common.dummy_aws_api import DummyAWSApi
+from tests.utils import MockedBoto3Request
 
 from .utils import assert_failure_messages
+
+
+@pytest.fixture()
+def boto3_stubber_path():
+    """Specify that boto3_mocker should stub calls to boto3 for the pcluster.utils module."""
+    return "pcluster.validators.awsbatch_validators.boto3"
 
 
 @pytest.mark.parametrize(
@@ -46,7 +61,11 @@ def test_awsbatch_region_validator(region, expected_message):
     ],
 )
 def test_compute_instance_type_validator(mocker, instance_type, max_vcpus, expected_message):
-    mocker.patch("pcluster.utils.get_supported_instance_types", return_value=["t2.micro", "p4d.24xlarge"])
+    mocker.patch("common.aws.aws_api.AWSApi.instance", return_value=DummyAWSApi())
+    mocker.patch(
+        "common.boto3.ec2.Ec2Client.describe_instance_type_offerings",
+        return_value=["t2.micro", "p4d.24xlarge"],
+    )
     mocker.patch(
         "pcluster.validators.awsbatch_validators.InstanceTypeInfo.init_from_instance_type",
         return_value=InstanceTypeInfo(
@@ -65,24 +84,9 @@ def test_compute_instance_type_validator(mocker, instance_type, max_vcpus, expec
     "min_vcpus, desired_vcpus, max_vcpus, expected_message",
     [
         (1, 2, 3, None),
-        (
-            3,
-            2,
-            3,
-            "desired vCPUs must be greater than or equal to min vCPUs",
-        ),
-        (
-            1,
-            4,
-            3,
-            "desired vCPUs must be fewer than or equal to max vCPUs",
-        ),
-        (
-            4,
-            4,
-            3,
-            "Max vCPUs must be greater than or equal to min vCPUs",
-        ),
+        (3, 2, 3, "desired vCPUs must be greater than or equal to min vCPUs"),
+        (1, 4, 3, "desired vCPUs must be fewer than or equal to max vCPUs"),
+        (4, 4, 3, "Max vCPUs must be greater than or equal to min vCPUs"),
     ],
 )
 def test_awsbatch_compute_resource_size_validator(min_vcpus, desired_vcpus, max_vcpus, expected_message):
@@ -119,7 +123,9 @@ def test_awsbatch_instances_architecture_compatibility_validator(
         return_value=[compute_architecture],
     )
     is_instance_type_patch = mocker.patch(
-        "pcluster.validators.awsbatch_validators.is_instance_type_format", side_effect=_internal_is_instance_type
+        "pcluster.validators.awsbatch_validators.AwsbatchInstancesArchitectureCompatibilityValidator."
+        "_is_instance_type_format",
+        side_effect=_internal_is_instance_type,
     )
 
     instance_types = compute_instance_types.split(",")
@@ -136,3 +142,231 @@ def test_awsbatch_instances_architecture_compatibility_validator(
     ]
     assert_that(supported_architectures_patch.call_count).is_equal_to(len(non_instance_families))
     assert_that(is_instance_type_patch.call_count).is_equal_to(len(instance_types))
+
+
+@pytest.mark.parametrize(
+    "raise_error_api_function, raise_error_parsing_function, types_parsed_from_emsg_are_known",
+    product([True, False], repeat=3),
+)
+def test_get_supported_batch_instance_types(
+    mocker, raise_error_api_function, raise_error_parsing_function, types_parsed_from_emsg_are_known
+):
+    """Verify functions are called and errors are handled as expected when getting supported batch instance types."""
+    # Dummy values
+    dummy_error_message = "dummy error message"
+    dummy_batch_instance_types = ["batch-instance-type"]
+    dummy_batch_instance_families = ["batch-instance-family"]
+    dummy_all_instance_types = dummy_batch_instance_types + ["non-batch-instance-type"]
+    dummy_all_instance_families = dummy_batch_instance_families + ["non-batch-instance-family"]
+    # Mock all functions called by the function
+    api_function_patch = mocker.patch(
+        "pcluster.validators.awsbatch_validators._get_cce_emsg_containing_supported_instance_types",
+        side_effect=BatchErrorMessageParsingException if raise_error_api_function else lambda: dummy_error_message,
+    )
+    parsing_function_patch = mocker.patch(
+        "pcluster.validators.awsbatch_validators._parse_supported_instance_types_and_families_from_cce_emsg",
+        side_effect=BatchErrorMessageParsingException
+        if raise_error_parsing_function
+        else lambda emsg: dummy_batch_instance_types + dummy_batch_instance_families,
+    )
+    mocker.patch("common.aws.aws_api.AWSApi.instance", return_value=DummyAWSApi())
+    supported_instance_types_patch = mocker.patch(
+        "common.boto3.ec2.Ec2Client.describe_instance_type_offerings",
+        return_value=dummy_all_instance_types,
+    )
+    get_instance_families_patch = mocker.patch(
+        "pcluster.validators.awsbatch_validators._get_instance_families_from_types",
+        return_value=dummy_all_instance_families,
+    )
+    candidates_are_supported_patch = mocker.patch(
+        "pcluster.validators.awsbatch_validators._batch_instance_types_and_families_are_supported",
+        return_value=types_parsed_from_emsg_are_known,
+    )
+    # this list contains values that are assumed to be supported instance types in all regions
+    assumed_supported = ["optimal"]
+    returned_value = awsbatch_validators._get_supported_batch_instance_types()
+    # The functions that call the Batch CreateComputeEnvironment API, and those that get all
+    # supported instance types and families should always be called.
+    assert_that(api_function_patch.call_count).is_equal_to(1)
+    assert_that(supported_instance_types_patch.call_count).is_equal_to(1)
+    get_instance_families_patch.assert_called_with(dummy_all_instance_types)
+    # The function that parses the error message returned by the Batch API is only called if the
+    # function that calls the API succeeds.
+    if raise_error_api_function:
+        parsing_function_patch.assert_not_called()
+    else:
+        parsing_function_patch.assert_called_with(dummy_error_message)
+    # The function that ensures the values parsed from the CCE error message are valid is only called if
+    # the API call and parsing succeed.
+    if any([raise_error_api_function, raise_error_parsing_function]):
+        candidates_are_supported_patch.assert_not_called()
+    else:
+        candidates_are_supported_patch.assert_called_with(
+            dummy_batch_instance_types + dummy_batch_instance_families,
+            dummy_all_instance_types + dummy_all_instance_families + assumed_supported,
+        )
+    # If either of the functions don't succeed, get_supported_instance_types return value should be
+    # used as a fallback.
+    assert_that(returned_value).is_equal_to(
+        dummy_all_instance_types + dummy_all_instance_families + assumed_supported
+        if any([raise_error_api_function, raise_error_parsing_function, not types_parsed_from_emsg_are_known])
+        else dummy_batch_instance_types + dummy_batch_instance_families
+    )
+
+
+@pytest.mark.parametrize(
+    "api_emsg, match_expected, expected_return_value",
+    [
+        # Positives
+        # Expected format using various instance types
+        ("be one of [u-6tb1.metal, i3en.metal-2tb]", True, ["u-6tb1.metal", "i3en.metal-2tb"]),
+        ("be one of [i3en.metal-2tb, u-6tb1.metal]", True, ["i3en.metal-2tb", "u-6tb1.metal"]),
+        ("be one of [c5n.xlarge, m6g.xlarge]", True, ["c5n.xlarge", "m6g.xlarge"]),
+        # Whitespace within the brackets shouldn't matter
+        ("be one of [c5.xlarge,m6g.xlarge]", True, ["c5.xlarge", "m6g.xlarge"]),
+        ("be one of [   c5.xlarge,     m6g.xlarge]", True, ["c5.xlarge", "m6g.xlarge"]),
+        ("be one of [c5.xlarge    ,m6g.xlarge    ]", True, ["c5.xlarge", "m6g.xlarge"]),
+        # Instance families as well as types must be handled.
+        ("be one of [  c5    ,    m6g  ]", True, ["c5", "m6g"]),
+        # The amount of whitespace between the words preceding the brackets doesn't matter.
+        ("be one of      [c5.xlarge, m6g.xlarge]", True, ["c5.xlarge", "m6g.xlarge"]),
+        ("be one of[c5.xlarge, m6g.xlarge]", True, ["c5.xlarge", "m6g.xlarge"]),
+        ("     be            one     of[c5.xlarge, m6g.xlarge]", True, ["c5.xlarge", "m6g.xlarge"]),
+        # Negatives
+        # Brackets are what's used to determine where the instance type list starts.
+        # If there are no brackets, there will be no match.
+        ("be one of (c5.xlarge, m6g.xlarge)", False, None),
+        ("be one of [c5.xlarge, m6g.xlarge", False, None),
+        ("be one of c5.xlarge, m6g.xlarge]", False, None),
+        # A comma must be used within the brackets.
+        ("be one of [c5.xlarge| m6g.xlarge]", False, None),
+    ],
+)
+def test_parse_supported_instance_types_and_families_from_cce_emsg(
+    caplog, api_emsg, match_expected, expected_return_value
+):
+    """Verify parsing supported instance types from the CreateComputeEnvironment error message works as expected."""
+    results_log_msg_preamble = "Parsed the following instance types and families from Batch CCE error message:"
+    caplog.set_level(logging.DEBUG)
+    if match_expected:
+        assert_that(
+            awsbatch_validators._parse_supported_instance_types_and_families_from_cce_emsg(api_emsg)
+        ).is_equal_to(expected_return_value)
+        assert_that(caplog.text).contains(results_log_msg_preamble)
+    else:
+        with pytest.raises(
+            BatchErrorMessageParsingException,
+            match="Could not parse supported instance types from the following: {0}".format(escape(api_emsg)),
+        ):
+            awsbatch_validators._parse_supported_instance_types_and_families_from_cce_emsg(api_emsg)
+        assert_that(caplog.text).does_not_contain(results_log_msg_preamble)
+
+
+@pytest.mark.parametrize("error_type", [ClientError, EndpointConnectionError(endpoint_url="dummy_endpoint"), None])
+def test_get_cce_emsg_containing_supported_instance_types(mocker, boto3_stubber, error_type):
+    """Verify CreateComputeEnvironment call to get error message with supported instance types behaves as expected."""
+    dummy_error_message = "dummy message"
+    call_api_patch = None
+    if error_type == ClientError:
+        mocked_requests = [
+            MockedBoto3Request(
+                method="create_compute_environment",
+                expected_params={
+                    "computeEnvironmentName": "dummy",
+                    "type": "MANAGED",
+                    "computeResources": {
+                        "type": "EC2",
+                        "minvCpus": 0,
+                        "maxvCpus": 0,
+                        "instanceTypes": ["p8.84xlarge"],
+                        "subnets": ["subnet-12345"],
+                        "securityGroupIds": ["sg-12345"],
+                        "instanceRole": "ecsInstanceRole",
+                    },
+                    "serviceRole": "AWSBatchServiceRole",
+                },
+                response=dummy_error_message,
+                generate_error=True,
+            )
+        ]
+        boto3_stubber("batch", mocked_requests)
+    else:
+        call_api_patch = mocker.patch(
+            "pcluster.validators.awsbatch_validators._call_create_compute_environment_with_bad_instance_type",
+            side_effect=error_type,
+        )
+
+    if error_type == ClientError:
+        return_value = awsbatch_validators._get_cce_emsg_containing_supported_instance_types()
+        assert_that(return_value).is_equal_to(dummy_error_message)
+    else:
+        expected_message = (
+            "Could not connect to the batch endpoint"
+            if isinstance(error_type, EndpointConnectionError)
+            else "did not result in an error as expected"
+        )
+        with pytest.raises(BatchErrorMessageParsingException, match=expected_message):
+            awsbatch_validators._get_cce_emsg_containing_supported_instance_types()
+        assert_that(call_api_patch.call_count).is_equal_to(1)
+
+
+@pytest.mark.parametrize(
+    "candidates, knowns",
+    [
+        (["c5.xlarge", "m6g"], ["c5.xlarge", "c5", "m6g.xlarge", "m6g"]),
+        (["bad-candidate"], ["c5.xlarge", "c5", "m6g.xlarge", "m6g"]),
+        (["optimal"], []),
+    ],
+)
+def test_batch_instance_types_and_families_are_supported(caplog, candidates, knowns):
+    """Verify function that describes whether all given instance types/families are supported behaves as expected."""
+    caplog.set_level(logging.DEBUG)
+    unknown_candidates = [candidate for candidate in candidates if candidate not in knowns]
+    expected_return_value = not unknown_candidates
+    observed_return_value = awsbatch_validators._batch_instance_types_and_families_are_supported(candidates, knowns)
+    assert_that(observed_return_value).is_equal_to(expected_return_value)
+    if unknown_candidates:
+        log_msg = "Found the following unknown instance types/families: {0}".format(" ".join(unknown_candidates))
+        assert_that(caplog.text).contains(log_msg)
+
+
+@pytest.mark.parametrize(
+    "instance_types, error_expected, expected_return_value",
+    [
+        (["m6g.xlarge"], False, ["m6g"]),
+        (["m6g.xlarge", "m6g-."], False, ["m6g", "m6g-"]),
+        (["m6g.xlarge", ".2xlarge"], True, ["m6g"]),
+    ],
+)
+def test_get_instance_families_from_types(caplog, instance_types, error_expected, expected_return_value):
+    """Verify the function that parses instance families from instance types works as expected."""
+    error_message_prefix = "Unable to parse instance family for instance type"
+    caplog.set_level(logging.DEBUG)
+    assert_that(awsbatch_validators._get_instance_families_from_types(instance_types)).contains_only(
+        *expected_return_value
+    )
+    if error_expected:
+        assert_that(caplog.text).contains(error_message_prefix)
+    else:
+        assert_that(caplog.text).does_not_contain(error_message_prefix)
+
+
+@pytest.mark.parametrize(
+    "candidate, expected_return_value",
+    [
+        ("m6g.xlarge", True),
+        ("c5n.xlarge", True),
+        ("i3en.metal-2tb", True),
+        ("u-6tb1.metal", True),
+        ("m6g", False),
+        ("c5n", False),
+        ("u-6tb1", False),
+        ("i3en", False),
+        ("optimal", False),
+    ],
+)
+def test_is_instance_type_format(candidate, expected_return_value):
+    """Verify function that decides whether or not a string represents an instance type behaves as expected."""
+    assert_that(AwsbatchInstancesArchitectureCompatibilityValidator._is_instance_type_format(candidate)).is_equal_to(
+        expected_return_value
+    )
